Initial implementation
Implements all of the EDL constructs supported by Ocaml oeedger8r except
- deep copy
- EDL validation, errors and warnings

For a given edl, that is valid and does not contain deep-copy annotations,
this implementation will produce the same output as the current oeedger8r.
The generated output files are of the form name_new.ext to distinguish from
and make it easy to compare against the output of oeedger8r.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>