--use-prefix feature.
When --use-prefix is supplied, then the edl name is prefixed to
all the untrusted proxies (ecall wrappers on the host side).
Ocalls are not affected.

This feature reduces the gap between OE SDK and Intel SDK.

Note:
The implementation  provides the ability to add prefixes to host-side ecall wrappers.
The ability to add prefixes to host side wrappers automatically can be leveraged for
a portable weak-symbol.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>