Nix based build shell for OE SDK.
oe-nix-shell provides a nix based build environment for OE SDK.
The build environment is reproducible. That is, on any machine that
supports nix package manager, the same versions of build tools will
be installed. The environment variables are also the same.
Irrespective of what else is installed in the system, the build
environment provided by oe-nix-shell is the same across all machines.

Though this PR tackles one of the core aspects of reproducible builds,
namely that the build environment is reproducible, it does not address
reproducible builds of the OE SDK. This is by design. Reproducible builds
of OE SDK can be attempted once the use of Nix has been vetted out in the
OE SDK's development and CI.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>